feat(common): add radix tree implementation for event/routing management
The implemented radix tree is general and global and is used to manage events and routers and the radix tree used for events and routers will be removed and both parts will use this tree. This tree is simpler than both the trees built for routers and events. It has O(1) and also supports parameters and wildcards. Also, there is no need for victor in Tree anymore.