feat(core): Unify event and routing systems with optimized radix tree implementation   (#77)
* feat(common): add radix tree implementation for event/routing management

The implemented radix tree is general and global and is used to manage events and routers and the radix tree used for events and routers will be removed and both parts will use this tree. This tree is simpler than both the trees built for routers and events. It has O(1) and also supports parameters and wildcards. Also, there is no need for victor in Tree anymore.

* refactor(events-http): Any radix-tree implementation for event/router has been removed and both parts use the Tree implementation. And with the changes synced, it still works as before. But better and more optimized and without the added overhead.