feat(core): The dependency-injection system was added.
This system uses http by default, but in the future, the Gland protocol will be supported.

This system is built completely similar to nestjs, only channel is used instead of provider. (Of course, it is weaker than nestjs). The items that have been added are:
- graph
- scanners
- compiler
- key-factory
- container
- app-initial
and the rest. These structures are the basic items for the release of the beta version of Gland.